refact: 마이페이지 레이아웃 서버 컴포넌트로 변경
